+++ spec/www/spec/db-sql-operations-test.js
@@ -1074,7 +1074,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/)
             else if (isAndroid && !isImpl2)
-              expect(error.message).toMatch(/syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/syntax error or other error.code: 1.message: near .true.: syntax error/);
             else
               expect(error.message).toMatch(/near \"true\": syntax error/);
 
@@ -1108,7 +1108,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
             else if (isAndroid && !isImpl2)
-              expect(error.message).toMatch(/syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/syntax error or other error.code: 1.message: near .false.: syntax error/);
             else
               expect(error.message).toMatch(/near \"false\": syntax error/);
 
@@ -1142,7 +1142,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
             else if (isAndroid && !isImpl2)
-              expect(error.message).toMatch(/syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/syntax error or other error.code: 1.message: near .Infinity.: syntax error/);
             else
               expect(error.message).toMatch(/near \"Infinity\": syntax error/);
 
@@ -1176,7 +1176,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
             else if (isAndroid && !isImpl2)
-              expect(error.message).toMatch(/syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/syntax error or other error.code: 1.message: near .-.: syntax error/);
             else
               expect(error.message).toMatch(/near \"-\": syntax error/);
 
@@ -1210,7 +1210,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
             else if (isAndroid && !isImpl2)
-              expect(error.message).toMatch(/syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/syntax error or other error.code: 1.message: near .NaN.: syntax error/);
             else
               expect(error.message).toMatch(/near \"NaN\": syntax error/);
 
@@ -1244,7 +1244,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
             else if (isAndroid && !isImpl2)
-              expect(error.message).toMatch(/syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/syntax error or other error.code: 1.message: near .SLCT.: syntax error/);
             else
               expect(error.message).toMatch(/near \"SLCT\": syntax error/);
 
@@ -1278,7 +1278,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
             else if (isAndroid && !isImpl2)
-              expect(error.message).toMatch(/syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/syntax error or other error.code: 1.message: near .SLCT.: syntax error/);
             else
               expect(error.message).toMatch(/near \"SLCT\": syntax error/);
 
@@ -1531,7 +1531,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
             else if (isAndroid && !isImpl2)
-              expect(error.message).toMatch(/syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/syntax error or other error.code: 1.message: near .SELCT.: syntax error/);
             else
               expect(error.message).toMatch(/near \"SELCT\": syntax error/);
 
@@ -1561,7 +1561,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/Error preparing an SQLite statement/);
             else if (isAndroid && !isImpl2)
-              expect(error.message).toMatch(/syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/syntax error or other error.code: 1.message: no such function: uper/);
             else
               expect(error.message).toMatch(/no such function: uper/);