+++ spec/www/spec/db-tx-error-handling-test.js
@@ -1762,7 +1762,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: SQLite3 step error result code: 21/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: other error.*code 21/);
+              expect(error.message).toMatch(/a statement with no error handler failed: other error.*code: 21 message: not an error/);
             else if (isAndroid && isImpl2)
               expect(error.message).toMatch(/a statement with no error handler failed: query not found/);
             else if (!isWebSql)
@@ -1830,7 +1830,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: SQLite3 step error result code: 21/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: other error.*code 21/);
+              expect(error.message).toMatch(/a statement with no error handler failed: other error.*code: 21 message: not an error/);
             else if (isAndroid && isImpl2)
               expect(error.message).toMatch(/a statement with no error handler failed: query not found/);
             else if (!isWebSql)
@@ -1898,7 +1898,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: SQLite3 step error result code: 21/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: other error.*code 21/);
+              expect(error.message).toMatch(/a statement with no error handler failed: other error.*code: 21 message: not an error/);
             else if (isAndroid && isImpl2)
               expect(error.message).toMatch(/a statement with no error handler failed: query not found/);
             else if (!isWebSql)
@@ -1967,7 +1967,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: SQLite3 step error result code: 21/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: other error.*code 21/);
+              expect(error.message).toMatch(/a statement with no error handler failed: other error.*code: 21 message: not an error/);
             else if (isAndroid && isImpl2)
               expect(error.message).toMatch(/a statement with no error handler failed: query not found/);
             else if (!isWebSql)
@@ -2036,7 +2036,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code: 1 message: near .*object Object.*: syntax error/);
             else if (!isWebSql)
               expect(error.message).toMatch(/a statement with no error handler failed:.*near \"\[object Object\]\": syntax error/);
             else if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
@@ -2106,7 +2106,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code: 1 message: near .*object Object.*: syntax error/);
             else if (!isWebSql)
               expect(error.message).toMatch(/a statement with no error handler failed:.*near \"\[object Object\]\": syntax error/);
             else if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
@@ -2176,7 +2176,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code: 1 message: near .*first.*: syntax error/);
             else if (!isWebSql)
               expect(error.message).toMatch(/a statement with no error handler failed:.*near \"first\": syntax error/);
             else if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
@@ -2242,7 +2242,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code: 1 message: near .*101.*: syntax error/);
             else if (!isWebSql)
               expect(error.message).toMatch(/a statement with no error handler failed:.*near \"101\": syntax error/);
             else if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
@@ -2309,7 +2309,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code: 1 message: near .*0.*: syntax error/);
             else if (!isWebSql)
               expect(error.message).toMatch(/a statement with no error handler failed:.*near \"0\": syntax error/);
             else if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
@@ -2506,7 +2506,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code: 1 message: near .*true.*: syntax error/);
             else if (!isWebSql)
               expect(error.message).toMatch(/a statement with no error handler failed:.*near \"true\": syntax error/);
             else if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
@@ -2572,7 +2572,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code: 1 message: near .*false.*: syntax error/);
             else if (!isWebSql)
               expect(error.message).toMatch(/a statement with no error handler failed:.*near \"false\": syntax error/);
             else if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
@@ -2638,7 +2638,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code: 1 message: near .Infinity.: syntax error/);
             else if (!isWebSql)
               expect(error.message).toMatch(/a statement with no error handler failed:.*near \"Infinity\": syntax error/);
             else if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
@@ -2703,7 +2703,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code: 1 message: near .-.: syntax error/);
             else if (!isWebSql)
               expect(error.message).toMatch(/a statement with no error handler failed:.*near \"-\": syntax error/);
             else if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
@@ -2768,7 +2768,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code: 1 message: near .NaN.: syntax error/);
             else if (!isWebSql)
               expect(error.message).toMatch(/a statement with no error handler failed:.*near \"NaN\": syntax error/);
             else if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
@@ -2833,7 +2833,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code: 1 message: near .NaN.: syntax error/);
             else if (!isWebSql)
               expect(error.message).toMatch(/a statement with no error handler failed:.*near \"NaN\": syntax error/);
             else if (isWebSql && !(/Android 4.[1-3]/.test(navigator.userAgent)))
@@ -3666,7 +3666,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code: 1 message: near .SLCT.: syntax error/);
             else if (!isWebSql)
               expect(error.message).toMatch(/a statement with no error handler failed.*near \"SLCT\": syntax error/);
             else
@@ -3725,7 +3725,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
             else if (!isWebSql && isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code: 1 message: near .SLCT.: syntax error/);
             else if (!isWebSql)
               expect(error.message).toMatch(/a statement with no error handler failed.*near \"SLCT\": syntax error/);
             else