+++ spec/www/spec/db-tx-error-mapping-test.js
@@ -85,7 +85,7 @@ var mytests = function() {
 
         // GENERAL NOTE: ERROR MESSAGES are subject to improvements and other possible changes.
 
-        it(suiteName + 'syntax error: command with misspelling [XXX MISSING ERROR DETAILS on Android ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40]', function(done) {
+        it(suiteName + 'syntax error: command with misspelling', function(done) {
           if (isWP8) pending('SKIP for WP(8)'); // SKIP for now
 
           var db = openDatabase("Syntax-error-test.db", "1.0", "Demo", DEFAULT_SIZE);
@@ -121,8 +121,7 @@ var mytests = function() {
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
               else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
-                // XXX MISSING ERROR DETAILS Android (Android-evcore-native-driver) ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40
-                expect(error.message).toMatch(/syntax error or other error.*code 1/);
+                expect(error.message).toMatch(/syntax error or other error code: 1 message: near .SLCT.: syntax error/);
               else if (!isWebSql && !isWindows && isAndroid && isImpl2)
                 expect(error.message).toMatch(/near \"SLCT\": syntax error.*code 1.*while compiling: SLCT 1/);
               else
@@ -163,7 +162,7 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + 'INSERT with VALUES in the wrong place (with a trailing space) [XXX TBD "incomplete input" vs "syntax error" message on (WebKit) Web SQL on Android 8.x/...; XXX MISSING ERROR DETAILS on Android ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40]', function(done) {
+        it(suiteName + 'INSERT with VALUES in the wrong place (with a trailing space) [XXX TBD "incomplete input" vs "syntax error" message on (WebKit) Web SQL on Android 8.x/...]', function(done) {
           if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
 
           var db = openDatabase("INSERT-Syntax-error-test.db", "1.0", "Demo", DEFAULT_SIZE);
@@ -206,7 +205,7 @@ var mytests = function() {
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
               else if (isAndroid && !isImpl2)
-                expect(error.message).toMatch(/syntax error or other error.*code 1/); // XXX MISSING ERROR DETAILS Android (Android-evcore-native-driver) ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40
+                expect(error.message).toMatch(/syntax error or other error code: 1 message: incomplete input/);
               else if (isAndroid && isImpl2)
                 expect(error.message).toMatch(/near \"VALUES\": syntax error.*code 1.*while compiling: INSERT INTO test_table/);
               else
@@ -235,9 +234,10 @@ var mytests = function() {
               expect(error.message).toMatch(/callback raised an exception.*or.*error callback did not return false/);
             else if (isWindows)
               expect(error.message).toMatch(/error callback did not return false.*Error preparing an SQLite statement/);
-            else if (!isWindows && isAndroid) // XXX TBD PROPER INFO MESSAGE NOT ON Android/...
-              // XXX MISSING ERROR DETAILS Android (Android-evcore-native-driver) ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40
-              expect(error.message).toMatch(/error callback did not return false.*syntax error/); // XXX Android/...
+            else if (isAndroid && !isImpl2)
+              expect(error.message).toMatch(/error callback did not return false.*syntax error or other error code: 1 message: incomplete input/);
+            else if (isAndroid && isImpl2)
+              expect(error.message).toMatch(/error callback did not return false.*syntax error/); // XXX TBD Android (built-in)
             else
               expect(error.message).toMatch(/error callback did not return false.*incomplete input/); // XXX SQLite 3.22.0 on iOS/macOS
 
@@ -302,7 +302,7 @@ var mytests = function() {
               else if (isWindows)
                 expect(error.message).toMatch(/SQLite3 step error result code: 1/);
               else if (isAndroid && !isImpl2)
-                expect(error.message).toMatch(/constraint fail.*code 19/);
+                expect(error.message).toMatch(/constraint fail.*code: 19 message: UNIQUE constraint failed: test_table\.data/);
               else if (isAndroid && isImpl2)
                 expect(error.message).toMatch(/constraint failure/);
               else
@@ -343,7 +343,7 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + 'SELECT uper("Test") (misspelled function name) [XXX MISSING ERROR DETAILS on Android ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40; INCORRECT error code WebKit Web SQL & plugin]', function(done) {
+        it(suiteName + 'SELECT uper("Test") (misspelled function name) [INCORRECT error code WebKit Web SQL & plugin]', function(done) {
           if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
 
           var db = openDatabase("Misspelled-function-name-error-test.db", "1.0", "Demo", DEFAULT_SIZE);
@@ -380,8 +380,7 @@ var mytests = function() {
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
               else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
-                // XXX MISSING ERROR DETAILS Android (Android-evcore-native-driver) ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40
-                expect(error.message).toMatch(/syntax error or other error.*code 1/);
+                expect(error.message).toMatch(/syntax error or other error.*code: 1 message: no such function: uper/);
               else if (!isWebSql && !isWindows && isAndroid && isImpl2)
                 expect(error.message).toMatch(/no such function: uper.*code 1/);
               else
@@ -412,7 +411,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/error callback did not return false.*Error preparing an SQLite statement/);
             else if (isAndroid && !isImpl2)
-              expect(error.message).toMatch(/error callback did not return false: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/error callback did not return false: syntax error or other error.*code: 1 message: no such function: uper/);
             else
               expect(error.message).toMatch(/error callback did not return false.*no such function: uper/);
 
@@ -424,7 +423,7 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + 'SELECT FROM bogus table (other database error) [XXX MISSING ERROR DETAILS on Android ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40; INCORRECT error code WebKit Web SQL & plugin]', function(done) {
+        it(suiteName + 'SELECT FROM bogus table (other database error) [INCORRECT error code WebKit Web SQL & plugin]', function(done) {
           if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
 
           var db = openDatabase("SELECT-FROM-bogus-table-error-test.db", "1.0", "Demo", DEFAULT_SIZE);
@@ -461,8 +460,7 @@ var mytests = function() {
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
               else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
-                // XXX MISSING ERROR DETAILS Android (Android-evcore-native-driver) ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40
-                expect(error.message).toMatch(/syntax error or other error.*code 1/);
+                expect(error.message).toMatch(/syntax error or other error code: 1 message: no such table: BogusTable/);
               else if (!isWebSql && !isWindows && isAndroid && isImpl2)
                 expect(error.message).toMatch(/no such table: BogusTable.*code 1/);
               else
@@ -492,7 +490,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/error callback did not return false.*Error preparing an SQLite statement/);
             else if (isAndroid && !isImpl2)
-              expect(error.message).toMatch(/error callback did not return false: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/error callback did not return false: syntax error or other error code: 1 message: no such table: BogusTable/);
             else
               expect(error.message).toMatch(/error callback did not return false.*no such table: BogusTable/);
 
@@ -504,7 +502,7 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + 'INSERT missing column [XXX MISSING ERROR DETAILS on Android ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40; INCORRECT error code WebKit Web SQL & plugin]', function(done) {
+        it(suiteName + 'INSERT missing column [INCORRECT error code WebKit Web SQL & plugin]', function(done) {
           if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
 
           var db = openDatabase("INSERT-missing-column-test.db", "1.0", "Demo", DEFAULT_SIZE);
@@ -542,9 +540,7 @@ var mytests = function() {
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
               else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
-                // XXX MISSING ERROR DETAILS Android (Android-evcore-native-driver) ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40
-                // expect(error.message).toMatch(/sqlite3_prepare_v2 failure:.*table test_table has 2 columns but 1 values were supplied/);
-                expect(error.message).toMatch(/syntax error or other error.*code 1/);
+                expect(error.message).toMatch(/code: 1 message: table test_table has 2 columns but 1 values were supplied/);
               else if (!isWebSql && !isWindows && isAndroid && isImpl2)
                 expect(error.message).toMatch(/table test_table has 2 columns but 1 values were supplied.*code 1.*while compiling: INSERT INTO test_table/);
               else
@@ -575,7 +571,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/error callback did not return false.*Error preparing an SQLite statement/);
             else if (isAndroid && !isImpl2)
-              expect(error.message).toMatch(/error callback did not return false: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/error callback did not return false: syntax error or other error code: 1 message: table test_table has 2 columns but 1 values were supplied/);
             else
               expect(error.message).toMatch(/error callback did not return false.*table test_table has 2 columns but 1 values were supplied/);
 
@@ -587,7 +583,7 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        it(suiteName + 'INSERT wrong column name [XXX MISSING ERROR DETAILS on Android ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40; INCORRECT error code WebKit Web SQL & plugin]', function(done) {
+        it(suiteName + 'INSERT wrong column name [INCORRECT error code WebKit Web SQL & plugin]', function(done) {
           if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
 
           var db = openDatabase("INSERT-wrong-column-name-test.db", "1.0", "Demo", DEFAULT_SIZE);
@@ -625,9 +621,7 @@ var mytests = function() {
               else if (isWindows)
                 expect(error.message).toMatch(/Error preparing an SQLite statement/);
               else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
-                // XXX MISSING ERROR DETAILS Android (Android-evcore-native-driver) ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40
-                // expect(error.message).toMatch(/sqlite3_prepare_v2 failure:.*table test_table has no column named wrong_column/);
-                expect(error.message).toMatch(/syntax error or other error.*code 1/);
+                expect(error.message).toMatch(/syntax error or other error.*code: 1 message: table test_table has no column named wrong_column/);
               else if (!isWebSql && !isWindows && isAndroid && isImpl2)
                 expect(error.message).toMatch(/table test_table has no column named wrong_column.*code 1.*while compiling: INSERT INTO test_table/);
               else
@@ -658,7 +652,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/error callback did not return false.*Error preparing an SQLite statement/);
             else if (isAndroid && !isImpl2)
-              expect(error.message).toMatch(/error callback did not return false: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/error callback did not return false: syntax error or other error.*code: 1 message: table test_table has no column named wrong_column/);
             else
               expect(error.message).toMatch(/error callback did not return false.*table test_table has no column named wrong_column/);
 
@@ -673,7 +667,7 @@ var mytests = function() {
         // NOTE: For some reason the Android/iOS (WebKit) Web SQL implementation
         // claims to detect the error at the "prepare statement" stage while the
         // plugin detects the error at the "execute statement" stage.
-        it(suiteName + 'CREATE VIRTUAL TABLE USING bogus module (other database error) [XXX MISSING ERROR DETAILS on Android ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40; INCORRECT error code WebKit Web SQL & plugin]', function(done) {
+        it(suiteName + 'CREATE VIRTUAL TABLE USING bogus module (other database error) [INCORRECT error code WebKit Web SQL & plugin]', function(done) {
           if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
 
           var db = openDatabase("create-virtual-table-using-bogus-module-error-test.db", "1.0", "Demo", DEFAULT_SIZE);
@@ -714,7 +708,7 @@ var mytests = function() {
               else if (isWindows)
                 expect(error.message).toMatch(/SQLite3 step error result code: 1/);
               else if (isAndroid && !isImpl2)
-                expect(error.message).toMatch(/syntax error or other error.*code 1/);
+                expect(error.message).toMatch(/syntax error or other error.*code: 1 message: no such module: bogus_module/);
               else if (isAndroid && isImpl2)
                 expect(error.message).toMatch(/no such module: bogus.*code 1/);
               else
@@ -744,7 +738,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/error callback did not return false.*SQLite3 step error result code: 1/);
             else if (isAndroid && !isImpl2)
-              expect(error.message).toMatch(/error callback did not return false: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/error callback did not return false: syntax error or other error.*code: 1 message: no such module: bogus_module/);
             else
               expect(error.message).toMatch(/error callback did not return false.*no such module: bogus/);
 
@@ -758,7 +752,7 @@ var mytests = function() {
 
         // TESTS with no SQL error handler:
 
-        it(suiteName + 'transaction.executeSql syntax error (command with misspelling) with no SQL error handler [XXX MISSING ERROR DETAILS on Android ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40]', function(done) {
+        it(suiteName + 'transaction.executeSql syntax error (command with misspelling) with no SQL error handler', function(done) {
           if (isWP8) pending('SKIP for WP(8)'); // FUTURE TBD
 
           db = openDatabase('tx-sql-syntax-error-with-no-sql-error-handler-test.db');
@@ -785,8 +779,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: Error preparing an SQLite statement/);
             else if (!isWebSql && !isWindows && isAndroid && !isImpl2)
-              // XXX MISSING ERROR DETAILS Android (Android-evcore-native-driver) ref: litehelpers/Cordova-sqlite-evcore-extbuild-free#40
-              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code 1/);
+              expect(error.message).toMatch(/a statement with no error handler failed: syntax error or other error.*code: 1 message: near .SLCT.: syntax error/);
             else if (!isWebSql && !isWindows && isAndroid && isImpl2)
               expect(error.message).toMatch(/a statement with no error handler failed: near \"SLCT\": syntax error.*code 1.*while compiling: SLCT 1/);
             else if (!isWebSql) // [iOS/macOS plugin]
@@ -842,7 +835,7 @@ var mytests = function() {
             else if (isWindows)
               expect(error.message).toMatch(/a statement with no error handler failed: SQLite3 step error result code: 1/);
             else if (isAndroid && !isImpl2)
-              expect(error.message).toMatch(/a statement with no error handler failed: constraint fail.*code 19/);
+              expect(error.message).toMatch(/a statement with no error handler failed: constraint fail.*code: 19 message: UNIQUE constraint failed: test_table.data/);
             else if (isAndroid && isImpl2)
               expect(error.message).toMatch(/a statement with no error handler failed:.*constraint failure/);
             else