+++ spec/www/spec/db-tx-value-bindings-test.js
@@ -1286,7 +1286,7 @@ var mytests = function() {
                 else if (isWindows)
                   expect(error.message).toMatch(/Error 25 when binding argument to SQL query/);
                 else if (isAndroid && !isImpl2)
-                  expect(error.message).toMatch(/other error.*code 25/);
+                  expect(error.message).toMatch(/other error.*code: 25 message: .*index.*out of range/);
                 else
                   expect(error.message).toMatch(/index.*out of range/);
 
@@ -1342,7 +1342,7 @@ var mytests = function() {
                 else if (isWindows)
                   expect(error.message).toMatch(/Error 25 when binding argument to SQL query/);
                 else if (isAndroid && !isImpl2)
-                  expect(error.message).toMatch(/other error.*code 25/);
+                  expect(error.message).toMatch(/other error.*code: 25 message: .*index.*out of range/);
                 else
                   expect(error.message).toMatch(/index.*out of range/);
 
@@ -1398,7 +1398,7 @@ var mytests = function() {
                 else if (isWindows)
                   expect(error.message).toMatch(/Error 25 when binding argument to SQL query/);
                 else if (isAndroid && !isImpl2)
-                  expect(error.message).toMatch(/other error.*code 25/);
+                  expect(error.message).toMatch(/other error.*code: 25 message: .*index.*out of range/);
                 else
                   expect(error.message).toMatch(/index.*out of range/);
 
@@ -1454,7 +1454,7 @@ var mytests = function() {
                 else if (isWindows)
                   expect(error.message).toMatch(/Error 25 when binding argument to SQL query/);
                 else if (isAndroid && !isImpl2)
-                  expect(error.message).toMatch(/other error.*code 25/);
+                  expect(error.message).toMatch(/other error.*code: 25 message: .*index.*out of range/);
                 else
                   expect(error.message).toMatch(/index.*out of range/);